Add query function to bq_utils

+++ data_steward/test/unit_test/bq_utils_test.py
@@ -1,12 +1,11 @@
 import unittest
 import time
-
 import bq_utils
 import gcs_utils
 from google.appengine.ext import testbed
-from test_util import FAKE_HPO_ID, FIVE_PERSONS_PERSON_CSV, NYC_FIVE_PERSONS_PERSON_CSV, PITT_FIVE_PERSONS_PERSON_CSV
-from test_util import NYC_FIVE_PERSONS_MEASUREMENT_CSV
+from test_util import FAKE_HPO_ID, FIVE_PERSONS_PERSON_CSV
 import resources
+import common
 
 
 PERSON = 'person'
@@ -39,7 +38,8 @@ class BqUtilsTest(unittest.TestCase):
 
     def _drop_created_test_tables(self):
         for table_id in self.tables_to_drop:
-            bq_utils.delete_table(table_id)
+            if table_id not in common.VOCABULARY_TABLES:
+                bq_utils.delete_table(table_id)
 
     def test_load_table_from_bucket(self):
         with open(FIVE_PERSONS_PERSON_CSV, 'rb') as fp:
@@ -48,8 +48,19 @@ class BqUtilsTest(unittest.TestCase):
         self.assertEqual(result['status']['state'], 'RUNNING')
 
     def test_load_table_from_bucket_error_on_bad_table_name(self):
-        with self.assertRaises(ValueError):
+        with self.assertRaises(ValueError) as cm:
             bq_utils.load_table_from_bucket(FAKE_HPO_ID, 'not_a_cdm_table')
+
+    def test_query_result(self):
+        with open(FIVE_PERSONS_PERSON_CSV, 'rb') as fp:
+            gcs_utils.upload_object(self.hpo_bucket, 'person.csv', fp)
+        bq_utils.load_table_from_bucket(FAKE_HPO_ID, PERSON)
+        time.sleep(2)
+
+        table_id = bq_utils.get_table_id(FAKE_HPO_ID, PERSON)
+        q = 'SELECT person_id FROM %s' % table_id
+        result = bq_utils.query(q)
+        self.assertEqual(5, int(result['totalRows']))
 
     def test_merge_with_good_data(self):
         with open(NYC_FIVE_PERSONS_PERSON_CSV, 'rb') as fp:
